==> Resume Parser & Classification (NLP)/1. NLP Project - Resume (Preprocessing - File Type Conversion).ipynb <==
### NLP Project - Resume (Preprocessing - File Type Conversion)

- In our Dataset of Resume we have 3 types of Files (.doc), (.docx) & (.pdf)
- We Need All the Files to be of one Extension i.e (.docx) for feature extraction / EDA
- Basically in this Step of Preprocessing we are making all the Files to have One Extension
 
#### Steps to follows for this file:
- Create a Folder "Resumes_Dataset" in your downloads folder.
- Copy all the 79 Resumes to this "Resumes_Dataset" Folder.
- Change the path "C:/Users/Akaash/Downloads/" to yours Accordingly and run the file Cell by Cell

#### Converting Doc to Docx

#!/usr/bin/env python
import os
import re
import glob
import win32com.client

word = win32com.client.Dispatch("Word.Application")

for i, doc in enumerate(glob.iglob("C:/Users/Akaash/Downloads/Resumes_Dataset/*.doc")):
    in_file = os.path.abspath(doc)
    wb = word.Documents.Open(in_file)
    out_file = re.sub(r'\.\w+$', '.docx', in_file)
    wb.SaveAs2(out_file, FileFormat=16) # file format for docx
    wb.Close()

word.Quit()

Inference: "The Resumes_Dataset" will have converted resumes to docx extension

#### Converting pdf to docx

#Importing the Required Library
from pdf2docx import parse
#Defining the Path
pdfdoc = 'C:/Users/Akaash/Downloads/Resumes_Dataset/Reactjs Developer_Prabakaran_Musquare Technologies.pdf'
todocx = 'C:/Users/Akaash/Downloads/Resumes_Dataset/Reactjs Developer_Prabakaran_Musquare Technologies.docx'
#Converting
parse(pdfdoc, todocx, start=0, end=None)

Inference: pdf to docx is converted

#### Deleting the Original Doc & pdf file

dir_name = "C:/Users/Akaash/Downloads/Resumes_Dataset/"
paths = os.listdir(dir_name)

for item in paths:
    if item.endswith(".doc"):
        os.remove(os.path.join(dir_name, item))
    elif item.endswith(".pdf"):
        os.remove(os.path.join(dir_name, item))

Inference: Original Files are Deleted, Now the "Resumes_Dataset" has 79 Files of (.docx) extension, will work will this Dataset for Further EDA
