==> Resume Parser & Classification (NLP)/3. NLP Project - Resume (Web Scrapping - List of Universities).ipynb <==
### NLP Project - Resume (Web Scrapping - List of Universities)

To Get List of Universities

# import module
import requests
from bs4 import BeautifulSoup

import warnings
warnings.filterwarnings("ignore")  #--to ignore warnings

HEADERS = ({'User-Agent':
            'Mozilla/5.0 (Windows NT 10.0; Win64; x64) \
            AppleWebKit/537.36 (KHTML, like Gecko) \
            Chrome/90.0.4430.212 Safari/537.36',
            'Accept-Language': 'en-US, en;q=0.5'})

# user define function
# Scrape the data
def getdata(url):
    r = requests.get(url, headers=HEADERS)
    return r.text


def html_code(url):

    # pass the url
    # into getdata function
    htmldata = getdata(url)
    soup = BeautifulSoup(htmldata, 'html.parser')

    #display html code
    return (soup)


url = "https://www.4icu.org/in/a-z/"

soup = html_code(url)

#Getting university Names

def uni_data(soup):
# find the Html tag
# with find()
# and convert into string
    data_str = ""
    uni_list = []

    for item in soup.find_all("a"):
        data_str = data_str + item.get_text()
        uni_list.append(data_str)
        data_str = ""
    return uni_list

#Listing the Universities
uni_res = uni_data(soup.body.tbody)

#To Save the Universities in CSV
import pandas as pd

# initialise data of lists.
data = {'University Name': uni_res,
        }

# Create DataFrame
df = pd.DataFrame(data)

# Save the output.
df.to_csv('C:/Users/Akaash/Downloads/List_of_universities.csv', index = None)
