==> Resume Parser & Classification (NLP)/4. NLP_Project_Resume_Classification_(Feature_Extraction).ipynb <==
### NLP Project - Resume Classification

#### Feature Extraction

##### Installed Required Library

#Uncomment & Run to Install
#!pip install pyresparser

#Importing Required Libraries
import numpy as np
import pandas as pd
import spacy
from spacy.matcher import Matcher
from os import listdir
from os.path import isfile, join
import re
#Downloading Stopwords
import nltk
nltk.download('stopwords')
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')
nltk.download('maxent_ne_chunker')
nltk.download('words')
from nltk.corpus import stopwords

#To Extract Skills
from pyresparser import ResumeParser
from nltk.tokenize import sent_tokenize

#Loading the csv Dataset that we got from Previous EDA Steps
df = pd.read_csv('/content/Resume_df.csv')
#Taking only the 'Text' & 'Clean Text' Column for Now
df = df.iloc[:,1:3]
#df = df.iloc[:,-1:]
df.head()

##### Name Extraction

# load pre-trained model
nlp = spacy.load('en_core_web_sm')
# initialize matcher with a vocab
matcher = Matcher(nlp.vocab)

#Function to Extract 'Name'
def extract_name(resume_text):
    nlp_text = nlp(resume_text)
    # First name and Last name are always Proper Nouns
    pattern = [{'POS': 'PROPN'}, {'POS': 'PROPN'}]
    matcher.add('NAME', [pattern], None)
    matches = matcher(nlp_text)
    for match_id, start, end in matches:
        span = nlp_text[start:end]
        return span.text

names = []
for i in range(len(df)):
  names.append(extract_name(df['Clean Text'][i]))

#Creating a 'Features' DataFrame
Features = pd.DataFrame(names,columns = ['Names'])
Features.head()

##### Mobile No Extraction

#Function To Extract Mobile Number
def extract_mobile_number(text):
    phone = re.findall(re.compile(r'(?:(?:\+?([1-9]|[0-9][0-9]|[0-9][0-9][0-9])\s*(?:[.-]\s*)?)?(?:\(\s*([2-9]1[02-9]|[2-9][02-8]1|[2-9][02-8][02-9])\s*\)|([0-9][1-9]|[0-9]1[02-9]|[2-9][02-8]1|[2-9][02-8][02-9]))\s*(?:[.-]\s*)?)?([2-9]1[02-9]|[2-9][02-9]1|[2-9][02-9]{2})\s*(?:[.-]\s*)?([0-9]{4})(?:\s*(?:#|x\.?|ext\.?|extension)\s*(\d+))?'), text)

    if phone:
        number = ''.join(phone[0])
        if len(number) > 10:
            return '+' + number
        else:
            return number

mobile = []
for i in range(len(df)):
  mobile.append(extract_mobile_number(df['Clean Text'][i]))

#Adding the "Mobile Number" Extracted
Features['Mobile_Number'] = mobile
Features.head()

##### Email ID Extraction

#Function To Extract Email ID
def extract_email(email):
    email = re.findall("[a-z0-9\.\-+_]+@[a-z0-9\.\-+_]+\.[a-z]+", email)
    if email:
        try:
            return email[0].split()[0].strip(';')
        except IndexError:
            return None

email = []
for i in range(len(df)):
  email.append(extract_email(df['Text'][i]))

#Adding the 'Email Id' Extracted
Features['Email'] = email
Features.head()

##### Skills Extraction

#Mounting Google Drive
from google.colab import drive
drive.mount('/content/drive')

#Getting Each Resumes Path
mypath = '/content/drive/MyDrive/Resumes_Dataset'
onlyfiles = [f for f in listdir(mypath) if isfile(join(mypath, f))]
#Using Resume Parser Library To Extract Skills
skills = []
for path in onlyfiles:
  skills.append(ResumeParser(f'/content/drive/My Drive/Resumes_Dataset/{path}').get_extracted_data()['skills'])

#Converting the List into an Array of List
skills = np.array(skills).reshape(-1,1)
#Converting into an DF
skills_df = pd.DataFrame(skills,columns = ['Skills'])
skills_df.head()

#Adding the "Skills" Extracted to Our "Features"
Features['Skills'] = skills_df['Skills']
Features.head()

##### Education Extraction

# load pre-trained model
nlp = spacy.load('en_core_web_sm')
# Grad all general stop words
STOPWORDS = set(stopwords.words('english'))
# Education Degrees
EDUCATION = [
            'BE','B.E.', 'B.E', 'BS','Bachelor of Technology','Senior Secondary' 'B.S', 'B.SC', 'B E', 'B. E.','B. E','B S','B. S','B. SC'
            'ME', 'M.E', 'M.E.', 'MS', 'M.S', 'B-TECH','M-TECH','M E', 'M. E','B.COM','B.ED','L.L.B.','LLB','LLM','L.L.M.', 'M. E.', 'M S', 'M. S',
            'BTECH', 'B.TECH', 'M.TECH', 'MTECH','B TECH', 'B. TECH', 'M. TECH', 'M TECH',
            'B. TECH','M. TECH','B TECH','M TECH','M.D.','NDA','N.D.A.','PHD','PGDM','P.G.D.M.' 'MBA','M.B.A.','MCA','M.C.A.','MS','M.S.','MD',
            'SSC', 'HSC', 'CBSE', 'ICSE', 'X', 'XII',
            'BBA','B.B.A.','BCA','B.C.A.','BA','B.A.', 
        ]


education = []
#Function to Extract Education
def extract_education(resume_text):
    nlp_text = nlp(resume_text)
    # Sentence Tokenizer
    nlp_text = [sent.text.strip() for sent in nlp_text.sents]

    edu = {}
    # Extract education degree
    for index, text in enumerate(nlp_text):
        for tex in text.split():
            # Replace all special symbols
            tex = re.sub(r'[?|$|.|!|,]', r'', tex)
            if tex.upper() in EDUCATION and tex not in STOPWORDS:
                edu[tex] = text + nlp_text[index]   

    # Extract year
    education = []
    if edu:
        for key in edu.keys():
            year = re.search(re.compile(r'(((20|19)(\d{2})))'), edu[key])  
            if year:
                education.append((key, ''.join(year[0])))
            else:
                education.append(key)
        return education
    else:
        EDU_PATTERN = re.findall(r'Bachelors? \D+|Masters? \D+',resume_text)   
        return EDU_PATTERN
        

# Calling function
for text in df['Text']:
    education.append(extract_education(text))

#Adding Education Extracted
Features['Education'] = education
Features.head()

##### Years of Experience Extraction

#Function to Extract Years of Experience
def extract_years_of_experience(text):
    lines = sent_tokenize(text)           
    experience = []
    for sentence in lines:
        if re.search('experience',sentence.lower()):        
            sen_tokenized = nltk.word_tokenize(sentence)
            tagged = nltk.pos_tag(sen_tokenized)            
            entities = nltk.chunk.ne_chunk(tagged)          
            for subtree in entities.subtrees():
                for leaf in subtree.leaves():
                    if leaf[1] == 'CD':                     
                        experience.append(leaf[0])
                        
    exp = []
    for ele in experience:
        if len(ele) <= 3 or (len(ele) <= 4 and ele[-1] == '0' 
                                and ele not in ('2020','2010','2000')):       
            exp.append(ele)
    if exp:
        return exp[0]
    else:
        return np.nan

years_of_experience = []
for i in range(len(df)):
    years_of_experience.append(extract_years_of_experience(df['Text'][i]))

#Adding "Years of Experience" Extracted
Features['Years_of_Experience'] = years_of_experience
Features.head()

##### Links Extraction 

links = []
def extract_links(text):
    x = re.findall(r'(https://(www.)?[a-z]+.com(/in)?/[a-z0-9]+)',text)   
    links = []
    for strings in x:
        for link in strings:
            if len(link) > 4:
                links.append(link)
    return links

for i in df['Text']:
    links.append(extract_links(text))

#Adding "Links" Extracted
Features['links'] = links
Features.head()

##### University Extraction

#Loading the csv which contains List of University
univ = pd.read_csv('/content/List_of_universities.csv')
univ.head()

#Funtion to get University Names
fun_uni = []

def get_uni_names(text): 
    University_Names = []
    lines = sent_tokenize(text)          
    for sentence in lines:                             
            if re.search('university',sentence.lower()) or re.search('education',sentence.lower()) or re.search('qualifications',sentence.lower()): # Search for words like education in sentence
                if re.search(r'\s{3,5}',sentence):  
                    sens = re.split(r'\s{3,5}', sentence)
                    for j in sens:
                        if re.search('university',j.lower()) or re.search('education',j.lower()) or re.search('qualifications',j.lower()):  # Search for keywords like education in splitted sentences
                            edu_sen = j
                            for k in univ['University Name']:  # Loop to run through each university name
                                if re.search(k.lower(),edu_sen.lower()):
                                    University_Names.append(k)   # If university is found, append it.
                                    

                else:                               # If there are no 3-5 spaces found, then run following
                    edu_sen = sentence
                    for k in univ['University Name']:      # Loop to run through each university name
                        if re.search(k.lower(),edu_sen.lower()):
                            University_Names.append(k)           # If university is found, append it.
            

                if University_Names:
                    continue
                else:
                    # print(sentence)
                    University_Names.append(sentence)    # If particular university name not found, append whole sentence with keywords like education.

    # Remove duplicates    
    for j in range(len(University_Names)-1):
        if University_Names[j] == University_Names[j+1]:
            del University_Names[j]
    
    return University_Names

for i in range(len(df)):
    fun_uni.append(get_uni_names(df['Text'][i]))

#Adding Universities Extracted
Features['Universities_Names'] = fun_uni
Features.head(10)

##### Saving the "Features" DF for Next Step of Model Building

Features.to_csv('Features.csv', index = None)
