==> Resume Parser & Classification (NLP)/5. NLP_Project_Resume_Classification_(Model_Building).ipynb <==
### NLP Project - Resume Classification

#### Model Building

# Mounting Google Drive
from google.colab import drive
drive.mount('/content/drive')

# Importing Libraries
import pandas as pd
import numpy as np
import os

#Models
from sklearn.linear_model import LogisticRegression
from sklearn.tree import  DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.multiclass import OneVsRestClassifier
from xgboost import XGBClassifier
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.naive_bayes import GaussianNB 

from sklearn.preprocessing import LabelEncoder
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split, KFold, GridSearchCV
from sklearn.metrics import accuracy_score, confusion_matrix

#NLP PreProcessing
import nltk
import re
import string
from nltk.stem import PorterStemmer, WordNetLemmatizer
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
nltk.download('stopwords')
nltk.download('punkt')
nltk.download('wordnet')

#Declaring the Classess Name Based on Our Resumes
ModelPath = '/content/drive/MyDrive/Original_Resumes'
Resume_Classes = ['Peoplesoft resumes','React resumes','SQL Developer Lightning insight','Workday resumes','Different resumes']

label = []
key = []
#Created a Function which will get a resume as 'key' & its coresponding class as 'labels'
def create_training_data():
    for category in Resume_Classes:
        path = os.path.join(ModelPath,category)
        class_num = Resume_Classes.index(category)
        for resume in os.listdir(path):
            label.append(category)
            key.append(resume)            

#Calling the Function
create_training_data()        

#Printing the Labels & Key First 10 Only.
print('Labels: \n{}\nLength of Labels: {}\nKeys: \n{}\nLength of Keys: {}'.format(label[:10], len(label), key[:10], len(key)))

# Creating a Dictionary of Key and Label (It will store Resumes as key & classs as value)
labelDict = dict(zip(key, label))

# Now Sorting it in alphabetical order of keys
finalDict = {}
for i in sorted(labelDict.keys()):
    finalDict[i] = labelDict[i]

# Let's Import Features.csv in which we had did feature extraction
Features = pd.read_csv('/content/Features.csv')
Features.head()

#Checking for NA Values
Features.isnull().sum()

# Creating a Dataframe for Modelling
Model_Features = pd.DataFrame(Features['Names'])
Model_Features['Skills'] = Features['Skills']
Model_Features['Label'] = labelDict.values()
Model_Features.head()

# Store the file into a csv file
Model_Features.to_csv('Model_features.csv', index = None)

#Loading the Dataset we created
df = pd.read_csv('/content/Model_features.csv')
df.head()

#Checking for NA Values
df.isnull().sum()

##### Data cleaning

#Initializing the Stop Words
lemmetizer = WordNetLemmatizer()
stop_words = stopwords.words('english')

#Function to get clean Skills
cleaned_data = []
def clean_data(text):
    text_clean = []
    text_tokens = word_tokenize(text)
    for word in text_tokens:
        if (word not in stop_words and # remove stopwords
            word not in string.punctuation): # remove punctuation
            stem_word = lemmetizer.lemmatize(word) # stemming word
            text_clean.append(stem_word)
    list_to_str = ' '.join([str(ele) for ele in text_clean])
    list_to_str = re.sub("'",'',list_to_str)   
    return list_to_str.lower() 

#Calling Function
for text in df['Skills']:
    cleaned_data.append(clean_data(text))

#Adding the Clean Skills to the Df
df['Cleaned Skills'] = cleaned_data
df.head()

##### Label Encoding

# Label Encoder
le = LabelEncoder()
df['Label'] = le.fit_transform(df['Label'])
df.head()

#Checking the Class value count
df['Label'].value_counts()

df.info()

# Store the file into a csv file
df.to_csv('Final_features.csv', index = None)

##### Train Test Split

#Taking only Skills & Label
x = df['Cleaned Skills'].values
y = df['Label'].values
#Splitting 80%-20%
x_train, x_test, y_train, y_test = train_test_split(x,y, test_size= 0.20, random_state= 38)
x_train.shape, x_test.shape, y_train.shape, y_test.shape

##### Model Building

#Vectoriation 
CV = CountVectorizer(max_features=20,stop_words = 'english')
x_train_cv = CV.fit_transform(x_train)
x_test_cv = CV.transform(x_test)

#Logistic regression and fit the model
classifier = LogisticRegression()
classifier.fit(x_train_cv,y_train)
classifier.score(x_test_cv, y_test)*100

##### Find best Algorithms using GridSearchCV

#Vectorizing x & y
x_cv = CV.fit_transform(x).toarray()

#GridSearchCV

def find_best_model_using_gridsearchcv(x,y):
    algos = {
        'Logistic' :{'Classifier': LogisticRegression(),'params':{'max_iter': [120,130,150]}
        },
        'Decision Tree': {'Classifier': DecisionTreeClassifier(),'params': {'criterion': ['gini','entropy'],'max_depth': [3,4,5]}
        },
        'Random Forest': {'Classifier': RandomForestClassifier(),'params':{'n_estimators':[100,125,150,175,200],'max_features': [4,5,6],'random_state':[7]}
        },
        'AdaBoost': {'Classifier': AdaBoostClassifier(),'params': {'n_estimators' : [10,15,20],'random_state':[7]}
        },
        'Gradient Boosting': {'Classifier': GradientBoostingClassifier(),'params': {'learning_rate':[0.1,0.01,0.2],'random_state':[7]}
        },
        'XGBM': {'Classifier': XGBClassifier(),'params': {'n_estimators' : [70,80,90,100],'max_depth': [3,4,5,7],
                                                          'learning_rate':[0.1,0.2],'random_state':[7]}
        },
        'SVM': {'Classifier': SVC(),'params': {'kernel':['rbf','poly'],'gamma':[50,100],'C':[10,15,20]}
        },
        'KNN': {'Classifier': KNeighborsClassifier(),'params': {'n_neighbors':[5,10,15,20]}
        },
        'MNB': {'Classifier': MultinomialNB(),'params': {'alpha':[0.5,1.0],'fit_prior':['True','False']}
        }
    }
    scores = []
    kfold = KFold(n_splits=5, random_state=10, shuffle=True)
    for algo_name, config in algos.items():
        gs =  GridSearchCV(config['Classifier'], config['params'], cv=kfold)
        gs.fit(x,y)
        scores.append({
            'Classifier': algo_name,
            'best_score': gs.best_score_,
            'best_params': gs.best_params_
        })

    return pd.DataFrame(scores,columns=['Classifier','best_score','best_params'])

find_best_model_using_gridsearchcv(x_cv,y)

Inference: The Best Algorithm we got from Grid Search is Decision Tree so will Create the final Model on Decision Tree Algorithms

##### Final Model (Decision Tree)

#Decision Tree Classifier and fit the model
DT_classifier = DecisionTreeClassifier(criterion = 'gini', max_depth = 5)
DT_classifier.fit(x_cv,y)
#Prediction Score
DT_classifier.score(x_cv,y)*100

#### Saving the Model

import pickle
filename = "resume.pickle"
pickle.dump(DT_classifier,open(filename,'wb'))

#Prediction
DT_classifier.predict(x_cv)

#Function to Get the Skills as a single list for Deployment
match = []
def match_words(text):
    lower = [x.lower() for x in text]

    for temp in lower:
        f_text = temp.split()
        found = False
        for text_word in f_text:
            match.append(text_word)
                                
#Calling the Function
match_words(df['Cleaned Skills'])

#Removing Duplicates From List 
final_skills = []
for i in match:
  if i not in final_skills:
    final_skills.append(i)

#Checking length of Final List
len(final_skills)

#Adding Data Science Skills Too.
DS = ['python','sql','excel','tableau','statistics','eda','feature engineering','machine learning','deep learning',
      'nlp','time series analysis','deployment']
final_skills.extend(DS)
len(final_skills)

#saving this skills in a text file to use for deployment
for skills in final_skills:
  with open("skills.txt", "a") as text_file:
      print(skills, file=text_file)
